Identify local and global optima.

Function that identifies local and global optima via variable neighborhood trust region search (VNTRS).

explore(f, npar, minimize = TRUE, controls = NULL, quiet = TRUE, seed = NULL)

Arguments

f
A function that computes value, gradient, and Hessian of the function to be optimized and returns them as a named list with elements value, gradient, and hessian. It is required that hessian is of class matrix.

npar
The number of parameters of f.

minimize
A boolean, determining whether f should be minimized minimize = TRUE (the default) or maximized minimize = FALSE.

controls
Either NULL or a named list with the following elements. Missing elements are set to default values (printed in parentheses).

init_runs (5): The number of short initial searches.

init_min (-1): The minimum value for the random initialization.

init_max (1): The maximum value for the random initialization.

init_iterlim (20): The number of iterations for the short initial searches.

neighborhoods (5): The number of nested neighborhoods.

neighbors (5): The number of neighbors in each neighborhood.

beta (0.05): A non-negative weight factor to account for the function's curvature in the selection of the neighbors. If beta = 0, the curvature is ignored. The higher the value, the higher probability of selecting a neighbor in the direction of the highest function curvature.

iterlim (1000): The maximum number of iterations to be performed before the local search is terminated.

only_global (FALSE): A boolean. Set only_global = TRUE if you are only interested in the global optimum. In this case, the algorithm will interrupt any local search prematurely that converges to a local optimum. This saves computation time. If only_global = FALSE (the default), the algorithm also looks for local optima.

tolerance (1e-5): Passed on to all.equal, which is used to check if two optima are equal.

quiet
Set quiet = FALSE to print progress.

seed
Set a seed for the sampling of the random starting points.

Value

A list of two list,

global optima and

local optima,

each of which contains lists with

value and

argument

of each identified optimum.

References

Bierlaire et al. (2009) "A Heuristic for Nonlinear Global Optimization" https://doi.org/10.1287/ijoc.1090.0343

Examples

### Gramacy & Lee function
gramacy_lee = function(x) {
 stopifnot(is.numeric(x))
 stopifnot(length(x) == 1)
 f = expression(sin(10*pi*x) / (2*x) + (x-1)^4)
 g = D(f, "x")
 h = D(g, "x")
 f = eval(f)
 g = eval(g)
 h = eval(h)
 list(value = f, gradient = g, hessian = as.matrix(h))
}
explore(f = gramacy_lee, npar = 1, seed = 1)

#> Error in explore(f = gramacy_lee, npar = 1, seed = 1): could not find function "explore"

### Shubert function
shubert = function(x) {
 stopifnot(is.numeric(x))
 stopifnot(length(x) == 2)
 f = expression(
   (cos(2*x1+1) + 2*cos(3*x1+2) + 3*cos(4*x1+3) + 4*cos(5*x1+4) +
   5*cos(6*x1+5)) *
   (cos(2*x2+1) + 2*cos(3*x2+2) + 3*cos(4*x2+3) + 4*cos(5*x2+4) +
   5*cos(6*x2+5))
 )
 g1 = D(f, "x1")
 g2 = D(f, "x2")
 h11 = D(g1, "x1")
 h12 = D(g1, "x2")
 h22 = D(g2, "x2")
 x1 = x[1]
 x2 = x[2]
 f = eval(f)
 g = c(eval(g1), eval(g2))
 h = rbind(c(eval(h11), eval(h12)), c(eval(h12), eval(h22)))
 list(value = f, gradient = g, hessian = h)
}
explore(f = shubert, npar = 2, seed = 1)

#> Error in explore(f = shubert, npar = 2, seed = 1): could not find function "explore"

### Rosenbrock function
rosenbrock = function(x) {
  stopifnot(is.numeric(x))
  stopifnot(length(x) == 2)
  f = expression(100 * (x2 - x1^2)^2 + (1 - x1)^2)
  g1 = D(f, "x1")
  g2 = D(f, "x2")
  h11 = D(g1, "x1")
  h12 = D(g1, "x2")
  h22 = D(g2, "x2")
  x1 = x[1]
  x2 = x[2]
  f = eval(f)
  g = c(eval(g1), eval(g2))
  h = rbind(c(eval(h11), eval(h12)), c(eval(h12), eval(h22)))
  list(value = f, gradient = g, hessian = h)
}
explore(f = rosenbrock, npar = 2, seed = 1)

#> Error in explore(f = rosenbrock, npar = 2, seed = 1): could not find function "explore"