Local trust region search

Function that links to trust.

Usage

local(f, parinit, minimize, controls, L)

Arguments

f

A function that computes value, gradient, and Hessian of the function to be optimized and returns them as a named list with elements value, gradient, and hessian.

parinit

Passed on to trust.

minimize

If TRUE, f gets minimized. If FALSE, maximized.

controls

Either NULL or a named list with the following elements. Missing elements are set to the default values in parentheses.

init_runs (5): The number of initial searches.

init_min (-1): The minimum argument value for the random initialization.

init_max (1): The maximum argument value for the random initialization.

init_iterlim (20): The number of iterations for the initial searches.

neighborhoods (5): The number of nested neighborhoods.

neighbors (5): The number of neighbors in each neighborhood.

beta (0.05): A non-negative weight factor to account for the function's curvature in the selection of the neighbors. If beta = 0, the curvature is ignored. The higher the value, the higher the probability of selecting a neighbor in the direction of the highest function curvature.

iterlim (1000): The maximum number of iterations to be performed before the local search is terminated.

tolerance (1e-6): A positive scalar giving the tolerance for comparing different optimal arguments for equality.

time_limit (NULL): The time limit in seconds for the algorithm.

L

A list of identified optima which contains lists with

value and

argument

of each identified optimum.

Value

A list of

success: A boolean, determining wether the local search successfully converged.

value: The value at the point where the local search terminated.

argument: The point where the local search terminated.