Select neighbors.

Function that selects neighbors around a given point x.

select_neighbors(f, x, neighborhood_expansion, controls)

Arguments

f
A function that computes value, gradient, and Hessian of the function to be optimized and returns them as a named list with elements value, gradient, and hessian. It is required that hessian is of class matrix.

x
A point in the domain of f.

neighborhood_expansion
A scaling factor, specifying the expansion of the neighborhood.

controls
Either NULL or a named list with the following elements. Missing elements are set to default values (printed in parentheses).

init_runs (5): The number of short initial searches.

init_min (-1): The minimum value for the random initialization.

init_max (1): The maximum value for the random initialization.

init_iterlim (20): The number of iterations for the short initial searches.

neighborhoods (5): The number of nested neighborhoods.

neighbors (5): The number of neighbors in each neighborhood.

beta (0.05): A non-negative weight factor to account for the function's curvature in the selection of the neighbors. If beta = 0, the curvature is ignored. The higher the value, the higher probability of selecting a neighbor in the direction of the highest function curvature.

iterlim (1000): The maximum number of iterations to be performed before the local search is terminated.

only_global (FALSE): A boolean. Set only_global = TRUE if you are only interested in the global optimum. In this case, the algorithm will interrupt any local search prematurely that converges to a local optimum. This saves computation time. If only_global = FALSE (the default), the algorithm also looks for local optima.

tolerance (1e-5): Passed on to all.equal, which is used to check if two optima are equal.

Value

A list points in the domain of f which neighbors of x.